+++ packages/docs-visual-areas/src/models/state.ts
@@ -1,5 +1,17 @@
 import * as vscode from 'vscode';
-import { getExtension } from './';
+import { extensionId } from './constants';
+
+export function getExtension() {
+	let extension: vscode.Extension<any> | undefined;
+	const ext = vscode.extensions.getExtension(extensionId);
+	if (!ext) {
+		throw new Error('Extension was not found.');
+	}
+	if (ext) {
+		extension = ext;
+	}
+	return extension;
+}
 
 export class State {
 	private static _extContext: vscode.ExtensionContext;
